Refactored chart functionality out of chart-ui
